BackButton Fix for iPad
Before that, the back button overlapped with the AirDrop button, it was difficult to press the back button, and the sensitivity field of the button was very small, so it was not always adjustable. Moved the button away from AirDrop and increased the field of vision. Changes made to the iPad screen.